Fusion Update
Added a new Fusion based Matte "Toggle" button to quickly invert the state of a matte entry in the "Matte List" field

Created a Fusion hotkeys file. Pressing the "Shift + A" hotkey in the Fusion Viewer window will toggle the active Cryptomatte "Matte Locator" selection state to switch between the "Add" and "Remove" modes.

Updated the README.md file with a Fusion screenshot and usage details.